Move common input classes to common folder
